+++ app/Core/Database/Migrations/DropMailLogIndexes.php
@@ -65,6 +65,8 @@ class DropMailLogIndexes extends AbstractMigration {
 				"<error>Migration $details failed: {$e->getMessage()}</error>"
 			);
 
+			$this->recordMigrationFailed();
+
 			return false;
 		}