+++ app/Core/Database/Migrations/MailLogDataMigration.php
@@ -63,6 +63,8 @@ class MailLogDataMigration extends AbstractMigration {
 				"<error>Migration $details failed: {$e->getMessage()}</error>"
 			);
 
+			$this->recordMigrationFailed();
+
 			return false;
 		}