+++ app/Core/Database/Migrations/MailRecipientsMigration.php
@@ -62,6 +62,8 @@ class MailRecipientsMigration extends AbstractMigration {
 				"<error>Migration $details failed: {$e->getMessage()}</error>"
 			);
 
+			$this->recordMigrationFailed();
+
 			return false;
 		}