+++ app/Core/Database/Migrations/MapUserConstraints.php
@@ -67,6 +67,8 @@ class MapUserConstraints extends AbstractMigration {
 				"<error>Migration $details failed: {$e->getMessage()}</error>"
 			);
 
+			$this->recordMigrationFailed();
+
 			return false;
 		}